Auth, validation

+++ routes/users.js
@@ -1,6 +1,12 @@
 const express = require('express');
-const userCtrl = require('../controllers/users');
-const { HttpBadRequestException } = require('../exceptions/httpExceptions');
+const auth = require('../middleware/auth');
+const { users: userCtrl } = require('../controllers');
+const userValidation = require('../utils/validation/userValidation');
+const transformJoiException = require('../utils/exceptions/transformJoiException');
+const { 
+    BadRequestException,
+    ForbidenException
+} = require('../utils/exceptions/UserFacingExceptions');
 
 const router = express.Router();
 
@@ -8,7 +14,13 @@ router.get('/:id', async (req, res, next) => {
     try {
         const { id } = req.params;
         if(!id) {
-            throw new HttpBadRequestException();
+            throw new BadRequestException();
+        }
+        const { error, value } = userValidation.validate({ 
+            id: id
+        });
+        if(error) {
+            transformJoiException(error, 'Unable to get user');
         }
         const user = await userCtrl.getUser(id);
         res.respData.data = user;
@@ -28,9 +40,17 @@ router.get('/', async (req, res, next) => {
 });
 router.post('/', async (req, res, next) => {
     try {
-        const { username, password } = req.body;
+        const { username, password, repeatPassword } = req.body;
         if(!username || !password) {
-            throw new HttpBadRequestException();
+            throw new BadRequestException();
+        }
+        const { error, value } = userValidation.validate({ 
+            username: username,
+            password: password,
+            repeatPassword: repeatPassword 
+        });
+        if(error) {
+            transformJoiException(error, 'Unable to create user');
         }
         const newUser = await userCtrl.addUser(username, password);
         res.respData.data = newUser;
@@ -40,11 +60,22 @@ router.post('/', async (req, res, next) => {
         next(error);
     }
 });
-router.patch('/', async (req, res, next) => {
+router.patch('/', auth, async (req, res, next) => {
     try {
         const { id, username, password } = req.body;
         if(!id) {
-            throw new HttpBadRequestException();
+            throw new BadRequestException();
+        }
+        if(req.user.id !== id) {
+            throw new ForbidenException();
+        }
+        const { error, value } = userValidation.validate({
+            id: id,
+            username: username,
+            password: password
+        });
+        if(error) {
+            transformJoiException(error, 'Unable to edit user');;
         }
         const editedUser = await userCtrl.editUser(id, username, password);
         res.respData.data = editedUser;
@@ -53,11 +84,20 @@ router.patch('/', async (req, res, next) => {
         next(error);
     }
 });
-router.delete('/', async (req, res, next) => {
+router.delete('/', auth, async (req, res, next) => {
     try {
         const { id } = req.body;
         if(!id) {
-            throw new HttpBadRequestException();
+            throw new BadRequestException();
+        }
+        if(req.user.id !== id) {
+            throw new ForbidenException();
+        }
+        const { error, value } = userValidation.validate({ 
+            id: id
+        });
+        if(error) {
+            transformJoiException(error, 'Unable to delete user');
         }
         await userCtrl.deleteUser(id);
         res.status(202);
